+++ src/routes/AppRoutes.tsx
@@ -5,13 +5,19 @@ import Content from "../pages/Content";
 import Blog from "../pages/blogs";
 import Video from "../pages/videos";
 import Article from "../pages/articles";
+import { useAppKitAccount } from "@reown/appkit/react";
+import { NotConnected } from "../components/ui/NotConnected";
 
 const AppRoutes = () => {
+  const { isConnected } = useAppKitAccount();
   return (
     <Router>
       <Routes>
         <Route path="/" element={<LandingPage />} />
-        <Route path="/dashboard" element={<Dashboard />} />
+        <Route
+          path="/dashboard"
+          element={isConnected ? <Dashboard /> : <NotConnected />}
+        />
         <Route path="/content" element={<Content />} />
         <Route path="/blogs" element={<Blog />} />
         <Route path="/videos" element={<Video />} />
